Update nteract elements: add inline renderer bundle injection
Support inline bundle injection via Vite's ?raw imports in IsolatedFrame
and OutputArea components from the nteract/elements registry. This eliminates
HTTP fetch overhead for the isolated renderer, enabling instant output rendering
in the notebook.

Key changes:
- Pull latest isolated-frame and output-area from registry via shadcn CLI
- Add rendererCode and rendererCss props to IsolatedFrame for inline bundle injection
- Update OutputArea to pass-through inline bundle props
- Import and pass pre-built bundle inline from CodeCell (zero HTTP requests)
- Add @isolated-bundle/ Vite alias for easy access to built bundles
- Update import paths from @/components/outputs/isolated to @/components/isolated
- Remove old duplicate isolated directory

This enables preloading the React renderer inline, making outputs render instantly
without waiting for async bundle fetch.
