--- src/components/outputs/isolated/__tests__/frame-html.test.ts
@@ -1,107 +1,0 @@
-/**
- * Tests for frame-html.ts HTML generation.
- *
- * These tests verify:
- * 1. Generated HTML has proper structure
- * 2. CSP meta tag is present
- * 3. Message handler validates source
- */
-
-import { describe, it, expect, beforeAll, beforeEach, afterEach, vi } from "vitest";
-import { createFrameBlobUrl, generateFrameHtml } from "../frame-html";
-
-describe("generateFrameHtml", () => {
-  let html: string;
-
-  beforeAll(() => {
-    html = generateFrameHtml({ darkMode: false });
-  });
-
-  it("generates valid HTML document", () => {
-    expect(html).toContain("<!DOCTYPE html>");
-    expect(html).toContain("<html");
-    expect(html).toContain("</html>");
-    expect(html).toContain("<head>");
-    expect(html).toContain("</head>");
-    expect(html).toContain("<body>");
-    expect(html).toContain("</body>");
-  });
-
-  it("includes Content-Security-Policy meta tag", () => {
-    expect(html).toContain('http-equiv="Content-Security-Policy"');
-  });
-
-  it("includes viewport meta tag", () => {
-    expect(html).toContain('name="viewport"');
-  });
-
-  /**
-   * SECURITY: Message handler must validate event.source.
-   * This prevents accepting messages from windows other than the parent.
-   */
-  it("message handler validates event.source", () => {
-    // The handler should check event.source === window.parent
-    expect(html).toContain("event.source !== window.parent");
-  });
-
-  it("sets up ready message listener", () => {
-    expect(html).toContain("addEventListener");
-    expect(html).toContain("message");
-  });
-
-  it("sends ready message on load", () => {
-    // The frame uses send('ready', null) which calls postMessage
-    expect(html).toContain("send('ready'");
-    expect(html).toContain("postMessage");
-  });
-
-  describe("dark mode", () => {
-    it("uses dark CSS variables when darkMode is true", () => {
-      const darkHtml = generateFrameHtml({ darkMode: true });
-      // Dark mode uses #0a0a0a for background
-      expect(darkHtml).toContain("--bg-primary: #0a0a0a");
-      expect(darkHtml).toContain("--text-primary: #e0e0e0");
-    });
-
-    it("uses light CSS variables when darkMode is false", () => {
-      const lightHtml = generateFrameHtml({ darkMode: false });
-      // Light mode uses #ffffff for background
-      expect(lightHtml).toContain("--bg-primary: #ffffff");
-      expect(lightHtml).toContain("--text-primary: #1a1a1a");
-    });
-  });
-});
-
-describe("createFrameBlobUrl", () => {
-  let mockCreateObjectURL: ReturnType<typeof vi.fn>;
-  let mockRevokeObjectURL: ReturnType<typeof vi.fn>;
-  let urlCounter = 0;
-
-  beforeEach(() => {
-    urlCounter = 0;
-    mockCreateObjectURL = vi.fn(() => `blob:mock-${++urlCounter}`);
-    mockRevokeObjectURL = vi.fn();
-    vi.stubGlobal("URL", {
-      ...URL,
-      createObjectURL: mockCreateObjectURL,
-      revokeObjectURL: mockRevokeObjectURL,
-    });
-  });
-
-  afterEach(() => {
-    vi.unstubAllGlobals();
-  });
-
-  it("returns a blob: URL", () => {
-    const url = createFrameBlobUrl({ darkMode: false });
-    expect(url).toMatch(/^blob:/);
-    expect(mockCreateObjectURL).toHaveBeenCalledWith(expect.any(Blob));
-  });
-
-  it("creates unique URLs each call", () => {
-    const url1 = createFrameBlobUrl({ darkMode: false });
-    const url2 = createFrameBlobUrl({ darkMode: false });
-    expect(url1).not.toBe(url2);
-    expect(mockCreateObjectURL).toHaveBeenCalledTimes(2);
-  });
-});
