--- src/components/outputs/isolated/__tests__/isolated-frame.test.ts
@@ -1,75 +1,0 @@
-/**
- * Tests for isolated-frame.tsx security invariants.
- *
- * CRITICAL SECURITY TESTS:
- * These tests verify the iframe sandbox configuration is secure.
- * If these tests fail, DO NOT PROCEED - the security model is broken.
- */
-
-import { describe, it, expect } from "vitest";
-
-/**
- * The sandbox attributes string from isolated-frame.tsx.
- * We duplicate it here to test against - if the source changes,
- * this test will catch discrepancies.
- */
-const EXPECTED_SANDBOX_ATTRS = [
-  "allow-scripts",
-  "allow-downloads",
-  "allow-forms",
-  "allow-pointer-lock",
-  "allow-popups",
-  "allow-popups-to-escape-sandbox",
-  "allow-modals",
-].join(" ");
-
-describe("iframe sandbox security", () => {
-  /**
-   * CRITICAL: The sandbox MUST NOT include allow-same-origin.
-   *
-   * If allow-same-origin is present, the iframe would:
-   * - Have access to the parent's origin
-   * - Be able to call Tauri APIs (window.__TAURI__)
-   * - Be able to access parent's DOM, cookies, localStorage
-   *
-   * This would completely break the security model.
-   */
-  it("sandbox does NOT include allow-same-origin", () => {
-    expect(EXPECTED_SANDBOX_ATTRS).not.toContain("allow-same-origin");
-  });
-
-  it("sandbox includes allow-scripts (required for widgets)", () => {
-    expect(EXPECTED_SANDBOX_ATTRS).toContain("allow-scripts");
-  });
-
-  it("sandbox includes allow-popups (required for links)", () => {
-    expect(EXPECTED_SANDBOX_ATTRS).toContain("allow-popups");
-  });
-
-  /**
-   * Verify we're not accidentally including dangerous permissions.
-   */
-  it("sandbox does NOT include allow-top-navigation", () => {
-    expect(EXPECTED_SANDBOX_ATTRS).not.toContain("allow-top-navigation");
-  });
-
-  it("sandbox does NOT include allow-top-navigation-by-user-activation", () => {
-    expect(EXPECTED_SANDBOX_ATTRS).not.toContain(
-      "allow-top-navigation-by-user-activation"
-    );
-  });
-});
-
-describe("sandbox attribute format", () => {
-  it("is a space-separated string", () => {
-    const parts = EXPECTED_SANDBOX_ATTRS.split(" ");
-    expect(parts.length).toBeGreaterThan(0);
-    // No empty parts (no double spaces)
-    expect(parts.every((p) => p.length > 0)).toBe(true);
-  });
-
-  it("all parts start with 'allow-'", () => {
-    const parts = EXPECTED_SANDBOX_ATTRS.split(" ");
-    expect(parts.every((p) => p.startsWith("allow-"))).toBe(true);
-  });
-});
