+++ src/components/outputs/media-router.tsx
@@ -1,18 +1,27 @@
 "use client";
 
-import { type ReactNode, Suspense } from "react";
+import { lazy, type ReactNode, Suspense } from "react";
 import { useMediaContext } from "./media-provider";
 
-// Direct imports for non-isolated output components (render in-DOM)
-// These are lightweight and don't need lazy loading
-import { AnsiOutput } from "./ansi-output";
-import { ImageOutput } from "./image-output";
-import { JsonOutput } from "./json-output";
-
-// NOTE: MarkdownOutput, HtmlOutput, and SvgOutput are intentionally NOT imported here.
-// These MIME types are always rendered in the isolated iframe (see OutputArea.tsx),
-// so including them in the main bundle would be wasteful duplication.
-// They are bundled separately in src/isolated-renderer/index.tsx.
+// Lazy load built-in output components for better bundle splitting
+const AnsiOutput = lazy(() =>
+  import("./ansi-output").then((m) => ({ default: m.AnsiOutput })),
+);
+const MarkdownOutput = lazy(() =>
+  import("./markdown-output").then((m) => ({ default: m.MarkdownOutput })),
+);
+const HtmlOutput = lazy(() =>
+  import("./html-output").then((m) => ({ default: m.HtmlOutput })),
+);
+const ImageOutput = lazy(() =>
+  import("./image-output").then((m) => ({ default: m.ImageOutput })),
+);
+const SvgOutput = lazy(() =>
+  import("./svg-output").then((m) => ({ default: m.SvgOutput })),
+);
+const JsonOutput = lazy(() =>
+  import("./json-output").then((m) => ({ default: m.JsonOutput })),
+);
 
 /**
  * Default MIME type priority order for rendering.
@@ -213,7 +222,7 @@ export function MediaRouter({
   metadata = {},
   priority: priorityProp,
   renderers: renderersProp,
-  unsafe: _unsafeProp, // Kept for API compatibility; HTML/Markdown now in IsolatedFrame
+  unsafe: unsafeProp,
   fallback,
   loading,
   className = "",
@@ -223,8 +232,7 @@ export function MediaRouter({
   // Props override context, context overrides built-in defaults
   const priority = priorityProp ?? ctx?.priority ?? DEFAULT_PRIORITY;
   const renderers = renderersProp ?? ctx?.renderers ?? {};
-  // Note: unsafe prop is kept for API compatibility with custom renderers,
-  // but built-in HTML/Markdown rendering is now handled by IsolatedFrame
+  const unsafe = unsafeProp ?? ctx?.unsafe ?? false;
 
   const mimeType = selectMimeType(data, priority);
 
@@ -256,31 +264,30 @@ export function MediaRouter({
   }
 
   const renderBuiltIn = () => {
-    // ISOLATED MIME TYPES - These are routed to IsolatedFrame by OutputArea.
-    // They should never reach MediaRouter in normal usage. If they do, it means
-    // OutputArea's isolation detection was bypassed (e.g., isolated={false}).
-    // Return null to avoid rendering potentially unsafe content in the main DOM.
-    if (
-      mimeType === "text/markdown" ||
-      mimeType === "text/html" ||
-      mimeType === "image/svg+xml"
-    ) {
-      if (process.env.NODE_ENV === "development") {
-        console.warn(
-          `MediaRouter received isolated MIME type "${mimeType}". ` +
-            `This should be rendered via IsolatedFrame instead.`
-        );
-      }
-      return null;
+    // Text/Markdown
+    if (mimeType === "text/markdown") {
+      return (
+        <MarkdownOutput
+          content={String(content)}
+          unsafe={unsafe}
+          className={className}
+        />
+      );
     }
 
-    // Plain text (may contain ANSI)
-    if (mimeType === "text/plain") {
-      return <AnsiOutput className={className}>{String(content)}</AnsiOutput>;
+    // HTML
+    if (mimeType === "text/html") {
+      return (
+        <HtmlOutput
+          content={String(content)}
+          unsafe={unsafe}
+          className={className}
+        />
+      );
     }
 
-    // Images (PNG, JPEG, GIF, WebP - NOT SVG which is isolated)
-    if (mimeType.startsWith("image/")) {
+    // Images (not SVG)
+    if (mimeType.startsWith("image/") && mimeType !== "image/svg+xml") {
       const imageType = mimeType as
         | "image/png"
         | "image/jpeg"
@@ -297,7 +304,12 @@ export function MediaRouter({
       );
     }
 
-    // JSON and structured data
+    // SVG
+    if (mimeType === "image/svg+xml") {
+      return <SvgOutput data={String(content)} className={className} />;
+    }
+
+    // JSON and structured data (but not custom +json types without a renderer)
     if (mimeType === "application/json") {
       return (
         <JsonOutput
@@ -306,6 +318,11 @@ export function MediaRouter({
           className={className}
         />
       );
+    }
+
+    // Plain text (may contain ANSI)
+    if (mimeType === "text/plain") {
+      return <AnsiOutput className={className}>{String(content)}</AnsiOutput>;
     }
 
     // Unknown +json types without custom renderer - show as JSON
@@ -323,8 +340,7 @@ export function MediaRouter({
     return <AnsiOutput className={className}>{String(content)}</AnsiOutput>;
   };
 
-  // Built-in components are eagerly loaded, no Suspense needed
-  return renderBuiltIn();
+  return <Suspense fallback={loadingComponent}>{renderBuiltIn()}</Suspense>;
 }
 
 /**
