+++ src/index.css
@@ -1,10 +1,36 @@
 @import "tailwindcss";
 @import "./styles/ansi.css";
 
+@custom-variant dark (&:is(.dark *));
+
 :root {
   --radius: 0.5rem;
   --background: 0 0% 100%;
   --foreground: 222.2 84% 4.9%;
+  --ansi-black: #3c3c3c;
+  --ansi-red: #c23621;
+  --ansi-green: #25bc24;
+  --ansi-yellow: #adad27;
+  --ansi-blue: #492ee1;
+  --ansi-magenta: #d338d3;
+  --ansi-cyan: #33bbc8;
+  --ansi-white: #cbcccd;
+  --ansi-bright-black: #818383;
+  --ansi-bright-red: #fc391f;
+  --ansi-bright-green: #31e722;
+  --ansi-bright-yellow: #eaec23;
+  --ansi-bright-blue: #5833ff;
+  --ansi-bright-magenta: #f935f8;
+  --ansi-bright-cyan: #14f0f0;
+  --ansi-bright-white: #e9ebeb;
+  --ansi-black-bg: #3c3c3c;
+  --ansi-red-bg: #c23621;
+  --ansi-green-bg: #25bc24;
+  --ansi-yellow-bg: #b5a000;
+  --ansi-blue-bg: #492ee1;
+  --ansi-magenta-bg: #d338d3;
+  --ansi-cyan-bg: #33bbc8;
+  --ansi-white-bg: #cbcccd;
 }
 
 @layer base {
@@ -14,4 +40,130 @@
   body {
     @apply bg-background text-foreground;
   }
+  .ansi-black-fg {
+    color: var(--ansi-black);
+  }
+  .ansi-red-fg {
+    color: var(--ansi-red);
+  }
+  .ansi-green-fg {
+    color: var(--ansi-green);
+  }
+  .ansi-yellow-fg {
+    color: var(--ansi-yellow);
+  }
+  .ansi-blue-fg {
+    color: var(--ansi-blue);
+  }
+  .ansi-magenta-fg {
+    color: var(--ansi-magenta);
+  }
+  .ansi-cyan-fg {
+    color: var(--ansi-cyan);
+  }
+  .ansi-white-fg {
+    color: var(--ansi-white);
+  }
+  .ansi-bright-black-fg {
+    color: var(--ansi-bright-black);
+  }
+  .ansi-bright-red-fg {
+    color: var(--ansi-bright-red);
+  }
+  .ansi-bright-green-fg {
+    color: var(--ansi-bright-green);
+  }
+  .ansi-bright-yellow-fg {
+    color: var(--ansi-bright-yellow);
+  }
+  .ansi-bright-blue-fg {
+    color: var(--ansi-bright-blue);
+  }
+  .ansi-bright-magenta-fg {
+    color: var(--ansi-bright-magenta);
+  }
+  .ansi-bright-cyan-fg {
+    color: var(--ansi-bright-cyan);
+  }
+  .ansi-bright-white-fg {
+    color: var(--ansi-bright-white);
+  }
+  .ansi-black-bg {
+    background-color: var(--ansi-black-bg);
+  }
+  .ansi-red-bg {
+    background-color: var(--ansi-red-bg);
+  }
+  .ansi-green-bg {
+    background-color: var(--ansi-green-bg);
+  }
+  .ansi-yellow-bg {
+    background-color: var(--ansi-yellow-bg);
+  }
+  .ansi-blue-bg {
+    background-color: var(--ansi-blue-bg);
+  }
+  .ansi-magenta-bg {
+    background-color: var(--ansi-magenta-bg);
+  }
+  .ansi-cyan-bg {
+    background-color: var(--ansi-cyan-bg);
+  }
+  .ansi-white-bg {
+    background-color: var(--ansi-white-bg);
+  }
+}
+
+.dark {
+  --ansi-black: #6e7681;
+  --ansi-red: #f97583;
+  --ansi-green: #85e89d;
+  --ansi-yellow: #ffea7f;
+  --ansi-blue: #79b8ff;
+  --ansi-magenta: #f692ce;
+  --ansi-cyan: #56d4dd;
+  --ansi-white: #fafbfc;
+  --ansi-bright-black: #959da5;
+  --ansi-bright-red: #fdaeb7;
+  --ansi-bright-green: #bef5cb;
+  --ansi-bright-yellow: #fff5b1;
+  --ansi-bright-blue: #c8e1ff;
+  --ansi-bright-magenta: #fedbf0;
+  --ansi-bright-cyan: #a9f1f7;
+  --ansi-bright-white: #ffffff;
+  --ansi-black-bg: #484f58;
+  --ansi-red-bg: #b31d28;
+  --ansi-green-bg: #176f2c;
+  --ansi-yellow-bg: #735c0f;
+  --ansi-blue-bg: #1158c7;
+  --ansi-magenta-bg: #5a32a3;
+  --ansi-cyan-bg: #0e7481;
+  --ansi-white-bg: #959da5;
+}
+
+@theme inline {
+  --color-ansi-black: var(--ansi-black);
+  --color-ansi-red: var(--ansi-red);
+  --color-ansi-green: var(--ansi-green);
+  --color-ansi-yellow: var(--ansi-yellow);
+  --color-ansi-blue: var(--ansi-blue);
+  --color-ansi-magenta: var(--ansi-magenta);
+  --color-ansi-cyan: var(--ansi-cyan);
+  --color-ansi-white: var(--ansi-white);
+  --color-ansi-bright-black: var(--ansi-bright-black);
+  --color-ansi-bright-red: var(--ansi-bright-red);
+  --color-ansi-bright-green: var(--ansi-bright-green);
+  --color-ansi-bright-yellow: var(--ansi-bright-yellow);
+  --color-ansi-bright-blue: var(--ansi-bright-blue);
+  --color-ansi-bright-magenta: var(--ansi-bright-magenta);
+  --color-ansi-bright-cyan: var(--ansi-bright-cyan);
+  --color-ansi-bright-white: var(--ansi-bright-white);
+  --color-ansi-black-bg: var(--ansi-black-bg);
+  --color-ansi-red-bg: var(--ansi-red-bg);
+  --color-ansi-green-bg: var(--ansi-green-bg);
+  --color-ansi-yellow-bg: var(--ansi-yellow-bg);
+  --color-ansi-blue-bg: var(--ansi-blue-bg);
+  --color-ansi-magenta-bg: var(--ansi-magenta-bg);
+  --color-ansi-cyan-bg: var(--ansi-cyan-bg);
+  --color-ansi-white-bg: var(--ansi-white-bg);
 }
